feat(FE-api-layer): commonQuery 분리 및 이에 따른 스키마 핸들러 쿼리 수정
